+++ app/Http/Controllers/ResumeController.php
@@ -188,19 +188,23 @@ class ResumeController extends Controller {// Клас по роботі з ре
         if(Auth::check())
         {
             $resume = $this->getResume($id);
+
             $cities = $city->getCities();
             $industries = $industry->getIndustries();
 
             $currency = new Currency();
             $currencies = $currency->getCurrencies();
 
-
+            if (User::find(Resume::find($resume->id)->id_u)->id==Auth::id())
             return view('Resume.edit')
                 ->with('resume',$resume)
                 ->with('cities',$cities)
                 ->with('industries',$industries)
                 ->with('currencies', $currencies);
+            else abort(403);
         }
+        else
+            return Redirect::to('auth/login');
     }
 
     /**
@@ -244,10 +248,12 @@ class ResumeController extends Controller {// Клас по роботі з ре
         {
             abort(500);
         }
-
-        Resume::destroy($id);
-
-        return redirect()->route('cabinet.index');
+        if (User::find(Resume::find($id)->id_u)->id==Auth::id()) {
+            Resume::destroy($id);
+            return redirect()->route('cabinet.index');
+        }
+        else
+            abort(403);
         //$resume->destroy();
     }
     public function send_message(Guard $auth,Request $request)
